Connecting news with BD
